retrieval-service/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
requests